Port to Minecraft 1.21.11
Toolchain: Gradle 9.6.1, Architectury Loom 1.17, architectury-plugin 3.5.
Dependencies: NeoForge 21.11.44, Fabric API 0.141.5, Architectury API 19.0.1,
Forge Config API Port 21.11.1 (v5 API), Mod Menu 17.0.0, Parchment 2025.12.20.

Code changes for vanilla API breaks since 1.21.1:
- ResourceLocation is named Identifier in current Mojang mappings
- jump key state moved to ClientInput.keyPresses (1.21.2 input rework)
- fall distance is a double in causeFallDamage (1.21.5)
- swim speed mixin retargets travelInWater after the travel() split
- LivingEntityInvoker removed; jumpFromGround is public now

+++ common/src/main/java/moe/qbyte/movement_plus/EntityHooks.java
@@ -19,7 +19,7 @@ public final class EntityHooks {
     }
 
     /**
-     * Scales the water-movement acceleration in {@code travel} for players;
+     * Scales the water-movement acceleration in {@code travelInWater} for players;
      * cross-loader replacement for NeoForge's {@code swim_speed} attribute.
      */
     public static float scaleSwimSpeed(LivingEntity entity, float amount) {
@@ -34,12 +34,12 @@ public final class EntityHooks {
      * The jump power compensation lives on the safe fall distance attribute; this only
      * covers midair jumps used since the last landing, which soften the fall like in 1.19.x.
      */
-    public static float modifyFallDistance(LivingEntity entity, float distance) {
+    public static double modifyFallDistance(LivingEntity entity, double distance) {
         if (!(entity instanceof Player player)) return distance;
 
         if (MidairJumpState.of(player).movement_plus$getUsedJumps() > 0) {
             double extraJumpHeight = JumpHeightHandler.extraJumpHeight(ServerConfig.jumpPowerMultiplier);
-            distance = (float) Math.max(0d, distance - extraJumpHeight / 1.5d - 1d);
+            distance = Math.max(0d, distance - extraJumpHeight / 1.5d - 1d);
         }
         return distance;
     }
